finished dar component
