+++ packages/bento-frontend/package.json
@@ -12,6 +12,7 @@
     "@bento-core/cart": "*",
     "@bento-core/tab": "*",
     "@bento-core/authentication": "*",
+    "@bento-core/data-access-request": "*",
     "@apollo/client": "^3.6.9",
     "@apollo/react-components": "^4.0.0",
     "@babel/core": "7.2.2",
