+++ packages/bento-frontend/src/pages/requestAccess/requestAccessController.js
@@ -3,14 +3,17 @@ import React from 'react';
 import { useQuery } from '@apollo/client';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import { Typography } from '../../components/Wrappers/Wrappers';
+import {DataAccessRequest} from '@bento-core/data-access-request';
+import { formFields, pageTitle, SUBMIT_REQUEST_ACCESS } from '../../bento/requestAccessData';
+import custodianUtils from '../../utils/custodianUtilFuncs';
+import Stats from '../../components/Stats/AllStatsController';
+import { bentoHelpEmail } from '../../bento/userLoginData';
 
 // Importing GraphQL Query.
 import {
   GET_ACCESS_CONTROL_LEVEL_DETAILS_QUERY,
 } from '../../bento/requestAccessData';
 
-// Importing View.
-import RequestAccessView from './requestAccessView';
 
 const accessRequestController = ({ match }) => {
   const { loading, error, data } = useQuery(GET_ACCESS_CONTROL_LEVEL_DETAILS_QUERY,
@@ -31,7 +34,15 @@ const accessRequestController = ({ match }) => {
 
   return (
     <div>
-      <RequestAccessView data={data} />
+      <Stats />
+      <DataAccessRequest 
+       data= {data} 
+       formFields = {formFields}
+       pageTitle = {pageTitle}
+       SUBMIT_REQUEST_ACCESS ={SUBMIT_REQUEST_ACCESS}
+       custodianUtils={custodianUtils}
+       bentoHelpEmail={bentoHelpEmail}
+      />
     </div>
   );
 };
