+++ packages/data-access-request/src/hooks/useQuery.js
@@ -1,0 +1,9 @@
+import React from "react";
+import { useLocation } from 'react-router-dom';
+
+function useQuery() {
+    const { search } = useLocation();
+    return React.useMemo(() => new URLSearchParams(search), [search]);
+  }
+
+  export default useQuery;
