+++ packages/data-access-request/src/index.js
@@ -1,0 +1,3 @@
+import DataAccessRequest from "./DataAccessRequest";
+
+export {DataAccessRequest};
