+++ packages/data-access-request/src/utils/getDefaultACL.js
@@ -1,0 +1,3 @@
+const getDefaultACL = (availableArms) => (availableArms[0] || []).id;
+
+export default getDefaultACL
