+++ packages/data-access-request/src/utils/getRedirectType.js
@@ -1,0 +1,6 @@
+function getRedirectedType(query) {
+    const path = query.get('type') || '/';
+    return path;
+  }
+
+  export default getRedirectedType;
