+++ packages/data-access-request/src/utils/isDisabledMember.js
@@ -1,0 +1,14 @@
+const isDisabledMember = (role, status) => {
+    const roles = [
+      'admin',
+      'member',
+    ];
+  
+    if (roles.includes(role.toLowerCase()) && status.toLowerCase === 'disabled') {
+      return false;
+    }
+  
+    return true;
+  };
+
+  export default isDisabledMember;
